feat(docs): render components as subpages only, split up examples and render them as tabs
