+++ mdx/link/cta.js
@@ -21,6 +21,7 @@ export const StyledCTA = styled(Link)(
  * Renders a Link styled as a call to action button.
  * @example
  * <CTA href="https://www.google.com/">CTA to google.com</CTA>
+ * @example
  * <CTA id="randomPageId" />
  */
 const CTA = (props) => <StyledCTA {...props} />
